helper methods for calling commands

+++ src/main/java/cz/bliksoft/meshcore/frames/cmd/CmdSetAdvertLatLon.java
@@ -17,7 +17,7 @@ public class CmdSetAdvertLatLon extends CommandFrame {
 	 * @param lon
 	 * @param alt
 	 */
-	public CmdSetAdvertLatLon(long lat, double lon, Integer alt) {
+	public CmdSetAdvertLatLon(double lat, double lon, Integer alt) {
 		this.lat = lat;
 		this.lon = lon;
 		this.alt = alt;
